+++ tests/tests_main.c
@@ -29,11 +29,11 @@ int main(int argc, char** argv)
     // registerTest("Video Test", videoTest);
     // registerTest("Monitor Test", monitorTest);
     // registerTest("Monitor Mode Test", monitorModeTest);
-    // registerTest("Window Test", windowTest);
+    registerTest("Window Test", windowTest);
     registerTest("Icon Test", iconTest);
-    // registerTest("Cursor Test", cursorTest);
-    // registerTest("Input Window Test", inputWindowTest);
-    // registerTest("System Cursor Test", systemCursorTest);
+    registerTest("Cursor Test", cursorTest);
+    registerTest("Input Window Test", inputWindowTest);
+    registerTest("System Cursor Test", systemCursorTest);
 #endif // PAL_HAS_VIDEO
 
 #if PAL_HAS_OPENGL